requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
packaging==24.0
PyMySQL==1.1.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
Werkzeug==3.0.1
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.27.0
pydantic==2.4.2
mysql-connector-python==8.3.0
motor==3.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0.post1
pymongo==4.6.1
